refactor(community): sign-in and sign-up forms update input components
